requirements.txt
fastapi
uvicorn
SQLAlchemy
PyMySQL
Werkzeug
Jinja2
python-dotenv
pydantic-settings